imagen

+++ Tema 2/CV/index.php
@@ -116,7 +116,9 @@ if (isset($_POST["boton_confirmar_editar"])) {
                             mysqli_query($conexion, $consulta);
 
                             if ($_POST["nombre_foto"] != "no_imagen.jpg") { //Si ya tenoia foto
-                                unlink("Img/" . $_POST["nombre_foto"]);
+                                if (is_file("Img/" . $_POST["nombre_foto"])){
+                                    unlink("Img/" . $_POST["nombre_foto"]);
+                                }                             
                             }
                         } catch (Exception $e) {
                             $mensaje = "Imposible subir la imagen. Error Nº " . mysqli_errno($conexion) . ": " . mysqli_error($conexion);
@@ -137,7 +139,6 @@ if (isset($_POST["boton_confirmar_editar"])) {
         }
     }
 }
-
 
 /***********************  NUEVO USUARIO - CONFIRMACIÓN   **********************/
 if (isset($_POST["boton_confirmar_nuevo"])) {
@@ -286,6 +287,25 @@ if (isset($_POST["boton_confirmar_nuevo"])) {
             border: none;
             background: none;
         }
+        .flexible{
+            display:flex;
+        }
+        .columna{
+            flex-flow: column;
+        }
+        .fila{
+            flex-flow: row;
+        }
+
+        div#edicion{
+            align-items: center;
+        }
+
+        div#editar_foto>img {
+            height: 15rem;
+            box-shadow: 0px 2px 10px 5px black;
+        }
+
     </style>
 </head>
 
@@ -478,75 +498,87 @@ if (isset($_POST["boton_confirmar_nuevo"])) {
 
 
     ?>
-        <form action="index.php" method="post" enctype="multipart/form-data" class="centrar">
-            <p>
-                <label for="nombre">Nombre:</label>
-                <input type="text" name="nombre" id="nombre" maxlength="30" value="<?php echo $nombre; ?>">
-                <?php if (isset($_POST["nombre"]) && $error_nombre) echo "<span class='error'> Campo vacío </span>" ?>
-            </p>
-            <p>
-                <label for="usuario">Usuario:</label>
-                <input type="text" name="usuario" id="usuario" maxlength="20" value="<?php echo $usuario; ?>">
-                <?php if (isset($_POST["usuario"]) && $error_usuario) {
-                    if ($_POST["usuario"] == "")
-                        echo "<span class='error'> Campo vacío </span>";
-                    else
-                        echo "<span class='error'> Usuario ya existente </span>";
-                }  ?>
-            </p>
-            <p>
-                <label for="clave">Contraseña:</label>
-                <input type="password" name="clave" id="clave" maxlength="20" placeholder="Editar contraseña">
-            </p>
+        <div id ="edicion" class="flexible fila">
 
-            <p>
-                <label for="dni">DNI:</label>
-                <input type="text" name="dni" id="dni" maxlength="10" value="<?php echo $dni; ?>" placeholder="DNI: 11223344Z">
-                <?php if (isset($_POST["dni"]) && $error_dni) {
-                    if ($_POST["dni"] == "")
-                        echo "<span class='error'> Campo vacío </span>";
-                    elseif (!dni_formato($_POST["dni"])) {
-                        echo "<span class='error'> Escriba un DNI con formato válido</span>";
-                    } elseif (!dni_valido($_POST["dni"]))
-                        echo "<span class='error'> DNI no válido</span>";
-                    else
-                        echo "<span class='error'>DNI ya en uso</span>";
-                } ?>
-            <p>
-                <label>Sexo:</label>
-                <?php if (isset($_POST["boton_confirmar_nuevo"]) && $error_sexo) echo "<span class='error'> Debe seleccionar un sexo </span>" ?>
-                <br />
-                <input type="radio" name="sexo" id="hombre" value="hombre" <?php if ($sexo == "hombre") echo "checked" ?>>
-                <label for="hombre"> Hombre</label><br />
-                <input type="radio" name="sexo" id="mujer" value="mujer" <?php if ($sexo == "mujer") echo "checked" ?>>
-                <label for="mujer"> Mujer</label>
-            </p>
 
-            <p>
-                <label for="foto">Incluir mi foto (max. 500KB) </label>
-                <input type="file" name="foto" id="foto" accept="image/*" />
+            <div>
+                <form action="index.php" method="post" enctype="multipart/form-data" class="centrar">
+                    <p>
+                        <label for="nombre">Nombre:</label>
+                        <input type="text" name="nombre" id="nombre" maxlength="30" value="<?php echo $nombre; ?>">
+                        <?php if (isset($_POST["nombre"]) && $error_nombre) echo "<span class='error'> Campo vacío </span>" ?>
+                    </p>
+                    <p>
+                        <label for="usuario">Usuario:</label>
+                        <input type="text" name="usuario" id="usuario" maxlength="20" value="<?php echo $usuario; ?>">
+                        <?php if (isset($_POST["usuario"]) && $error_usuario) {
+                            if ($_POST["usuario"] == "")
+                                echo "<span class='error'> Campo vacío </span>";
+                            else
+                                echo "<span class='error'> Usuario ya existente </span>";
+                        }  ?>
+                    </p>
+                    <p>
+                        <label for="clave">Contraseña:</label>
+                        <input type="password" name="clave" id="clave" maxlength="20" placeholder="Editar contraseña">
+                    </p>
+
+                    <p>
+                        <label for="dni">DNI:</label>
+                        <input type="text" name="dni" id="dni" maxlength="10" value="<?php echo $dni; ?>" placeholder="DNI: 11223344Z">
+                        <?php if (isset($_POST["dni"]) && $error_dni) {
+                            if ($_POST["dni"] == "")
+                                echo "<span class='error'> Campo vacío </span>";
+                            elseif (!dni_formato($_POST["dni"])) {
+                                echo "<span class='error'> Escriba un DNI con formato válido</span>";
+                            } elseif (!dni_valido($_POST["dni"]))
+                                echo "<span class='error'> DNI no válido</span>";
+                            else
+                                echo "<span class='error'>DNI ya en uso</span>";
+                        } ?>
+                    <p>
+                        <label>Sexo:</label>
+                        <?php if (isset($_POST["boton_confirmar_nuevo"]) && $error_sexo) echo "<span class='error'> Debe seleccionar un sexo </span>" ?>
+                        <br />
+                        <input type="radio" name="sexo" id="hombre" value="hombre" <?php if ($sexo == "hombre") echo "checked" ?>>
+                        <label for="hombre"> Hombre</label><br />
+                        <input type="radio" name="sexo" id="mujer" value="mujer" <?php if ($sexo == "mujer") echo "checked" ?>>
+                        <label for="mujer"> Mujer</label>
+                    </p>
+
+                    <p>
+                        <label for="foto">Incluir mi foto (max. 500KB) </label>
+                        <input type="file" name="foto" id="foto" accept="image/*" />
+                        <?php
+                        if (isset($_POST["boton_confirma_nuevo"]) && $error_foto) {
+
+                            if ($_FILES["foto"]["error"]) {
+                                echo "<span class='error'>Error en la subida del fichero</span>";
+                            } elseif (!getimagesize($_FILES["foto"]["tmp_name"])) {
+                                echo "<span class='error'>El archivo debe ser una imagen</span>";
+                            } else {
+                                echo "<span class='error'>La imagen no debe exceder los 500KB</span>";
+                            }
+                        }
+                        ?>
+                    </p>
+
+                    <p>
+                        <button type="submit" name="boton_volver">Volver</button>
+                        <input type="hidden" name="nombre_foto" value="<?php echo $foto; ?>">
+                        <!--input type="hidden" name ="id_usuario" value = ""-->
+                        <button type="submit" name="boton_confirmar_editar" value="<?php echo $id_usuario; ?>">Guardar cambios</button>
+                    </p>
+            </div>
+            <div class="flexible columna" id ="editar_foto">
+                <img src="Img/<?php echo $foto; ?>" alt="Foto de perfil" title="Foto de perfil" />
                 <?php
-                if (isset($_POST["boton_confirma_nuevo"]) && $error_foto) {
-
-                    if ($_FILES["foto"]["error"]) {
-                        echo "<span class='error'>Error en la subida del fichero</span>";
-                    } elseif (!getimagesize($_FILES["foto"]["tmp_name"])) {
-                        echo "<span class='error'>El archivo debe ser una imagen</span>";
-                    } else {
-                        echo "<span class='error'>La imagen no debe exceder los 500KB</span>";
-                    }
-                }
+                if ($foto != "no_imagen.jpg")
+                    echo "<p class='texto-centrado'><button type='submit' name='borrar_foto' value='".$foto."'>Borrar</button></p>"
                 ?>
-            </p>
-
-            <p>
-                <button type="submit" name="boton_volver">Volver</button>
-                <input type="hidden" name="nombre_foto" value="<?php echo $foto; ?>">
-                <!--input type="hidden" name ="id_usuario" value = ""-->
-                <button type="submit" name="boton_confirmar_editar" value="<?php echo $id_usuario; ?>">Guardar cambios</button>
-            </p>
-        </form>
-
+            </div>
+            </form>
+        </div>
     <?php
 
     }
@@ -605,6 +637,7 @@ if (isset($_POST["boton_confirmar_nuevo"])) {
         mysqli_close($conexion);
         die($mensaje);
     }
+
 
     ?>
 </body>
